Fix plugin hanging on simple content-types when caching is enabled
A simple content-type will have an empty resolved populate object which should be passed as a `null` when resolving, otherwise a non-breaking loop will occur.

This is a prime example on why we need to add end-to-end tests that actually call the playground API, as this bug was part of the documents hook that is hard to trigger from unit or integration tests.

Fixes #52